Adding new apache cookbook.
This isn't complete yet - but I want to get it out there so that (1)
people know I'm working on this (this came from my work on SCALE stuff)
and don't duplicate effort and (2) people can start using it if they
want.

It was primarily written as part of my SCALE work, and works on Debian,
but I'll cleanup the RH portions after I merge it.