send syslog messages to scribe directly from hosts
Summary:
We (security) currently aggregate syslog messages from every prod machine to central syslog servers (as defined by node.get_svc_ipaddr('syslog', false)) and then do nothing else with those messages but put them into a scribe category. We chose to do this because we used to do extra processing of those messages on the central syslog hosts, but we don't do that anymore and simply drop them into scribe using the program output module in rsyslog (omprog).

I would like to get rid of centrally collecting syslog messages in favor of sending messages to scribe from each individual host. This would eliminate the need to have (and maintain and monitor) central syslog servers and the need to have every server send syslog over the network. We would use the exact same method that we use on central syslog servers on every server to send messages to scribe.

This diff implements the logic to send syslog messages into scribe directly from each machine using the program output module in rsyslog. We will shard it out and monitor for issues.

The binary that is used to send syslog to scribe is very simple. It takes stdin and put it into scribe. On central syslog hosts it takes up 10MB of RSS, and has memory checkers that kills it if it goes over 30MB. Here it is running on my dev server

root     2175569  0.0  0.0  77548 10308 ?        Sl   16:28   0:00 [omscribe]

This also enables debug logging for rsyslogd so that we can keep track of queues on hosts, since now that our only place to monitor rsyslogd is on each machine vs central hosts.

Reviewed By: joshuamiller01

Differential Revision: D3603289

fbshipit-source-id: 9c0f1571af25d4329388e0e9865bbdcec1699f14