Add an optional fallback to rmmod when modprobe fails
Summary:
This diff was primarily to do the other thing, which is:
```
  if node['mdadm'] && node['mdadm'][raidname]
```
but then I realized that, if one would `yum erase nvidia*`, then chef would fail as it might try to unload the `nvidia` module using `modprobe`, which fails as `nvidia` isn't a module in `modules.dep`.
So falling back on `rmmod` sounded like a great idea. jcalvinowens didn't really want everything to use that, so I made it optional.

Reviewed By: jaymzh

Differential Revision: D4180296

fbshipit-source-id: d1507b38b84012f94b0805a976d6f4eee67ea331