fb_fstab: don't skip /dev/fioa* when parsing initial fstab
Summary:
The CentOS 7 udb I was testing D4260763 on kept coming up
with / in ro mode. Turns out /etc/fstab was empty on this host (side note, this
is probably one of the reasons dracut was breaking things.) The chef.first.out
skipped all the /dev/fioa* entries since it wasn't in the whitelist of entries
to keep.

Reviewed By: bwann

Differential Revision: D4261436

fbshipit-source-id: 8fd156571e2870f3f51f296cf6b95b3234a19d79