Introduce /var/facebook/fb_grub.use_labels file and move all new grub 2 based installations over to label based grub
Summary:
The general idea: There is no good reason why we should deal with all of the guessing games we do.
My last diff made it slightly less painfull, but there are still cases where even that doesn't work.
Labels is the future and I personally welcome our new label overlords.

This diff is yet another "we want to roll it out to everything from here on out" kind of deal.
Files seem to work ok with that.
malmond77 would also be able to use that in his hotmetal provisioning runs.

There is a 2nd diff that will drop that file i place in the anaconda post script; D4622991

On those machines we will be able to remove the grub.before_chef file as well.

Reviewed By: odcinek

Differential Revision: D4615438

fbshipit-source-id: 39f8287127e873ddd4141b765f585c562cf32431