Enable label based grub for all centos7 LEOPARD_BDW_T4 boxes
Summary:
There is now support for using disk labels to identify the boot
partition in grub configuration, but this is disabled by default.

This change enables support on all centos7 machines with model_name
LEOPARD_BDW_T4 since these machines are currently failing to boot
with the default BIOS settings as detailed in the task referenced
below.

Reviewed By: pallotron

Differential Revision: D4642521

fbshipit-source-id: bb197c66c5c8d5a508028f45100f191a1eba5638