Next try: Make sure that EFI and non-EFI packages and configs of grub2 are available
Summary:
Previously: D4702750
Reverted: D4718229

The only change to the previous one is that I clean up the existing pacakges first, then install the new ones.
During the anaconda install, we do install about half of them. We always grab the fb2 version since we don't specify.
When we then want to install the rest, it's a 50% upgrade/install, 50% downgrade. That's something the yum provider doesn't support. That's why it broke the last time.

Reviewed By: joshuamiller01

Differential Revision: D4766014

fbshipit-source-id: 0dce0546e1f8483e2fb4b388c0bb7f93cd357df8