Correctly write EFI and BIOS grub2 configs respectively
Summary:
We did write out both, EFI and BIOS configuration files. However: In the BIOS configuration file, we still wrote EFI values :(

As far as the code goes:
Turns out that variable was literally only used in a single template. I moved it locally and made it respect the possible bios/efi options.

Reviewed By: jaymzh

Differential Revision: D4780949

fbshipit-source-id: 0f27c3b13b6cd7395395d08d087ff2d1632ffd29