add new tboot cmdline arg for failsafe
Summary: Sahil recently made a diff to tboot (see D4724533 and D4933408) so that if there was a TXT error on a previous boot (for example, Vt-D is not enabled) then tboot will fallthrough to launching the kernel without trying to do GETSEC[ENTER] again.  For backwards compatibility, that feature was added as an option which I'm now adding to our local configs here.

Reviewed By: jaymzh

Differential Revision: D4934572

fbshipit-source-id: e87ca76a76b6ca5b80d0475a6c0ae863293e8b81