fb_timer user jobs
Summary:
We need the ability to run systemd user timers on shell servers.  Currently all fb_timers are created as system units.

This diff adds a flag (delay_start) to the timer node to delay starting each timer during the chef run.  This, combined with setting the unit path to the /user directory within /etc/systemd allows using fb_timers to create user jobs.  The user jobs can be started and stopped by other means.

This is intended as a fix for t16940241

Reviewed By: jaymzh, zfjagann

Differential Revision: D4876335

fbshipit-source-id: 5a8dc690bc1fbc67fe08b8969fe8a7dc1002b3e8