use true tboot-1.9.5 grub style
Summary:
When I migrated tboot into yummy I imported the most recent code from the sourceforge repo. That turned out to have a security fix which was not backwards compatible and I backed out temporarily.

In the parent task I put together all the changes needed to make the new version work in our environment, and this is the final step of fixing the grub config to match the input that tboot is now expecting.

Reviewed By: joshuamiller01

Differential Revision: D4960486

fbshipit-source-id: cfc8caa251bb3ac991eafe562c0005ce51759eb6