Handle sysctls idempotently
Summary:
We were relying on `sysctl -p` to be idempotent (we only ran it when we knew
*something* had changed, but in those cases it set *every* setting, ew). This
causes problems when setting certain sysctls that - even when the same value -
have side effects upon writing them.

This moves the checking and convergance logic into Chef itself. It also adds
unit tests for that, and we're sharding the new logic out.

Reviewed By: joshuamiller01

Differential Revision: D5198360

fbshipit-source-id: 88e28ba30fca1656aee8883ea778ee4e37a51cd2