adding host_local_entries to disable default hosts entries
Summary:
```

```

In corp, oracle have requirement to not to have local /etc/hosts entries, Here is the explaination from Oracle team

```
For Oracle Applications Disaster Recovery the config should have the same hostname as Primary host, For this we have masked the /etc/hosts to have the primary hostname resolving to DR site hosts, This is standard Oracle requirement for Disaster Recovery using host file virtual names.

In case of using fbit_host in chef, It is populating the target machines IP's address (which is fine in the normal case), But in our case it is getting the hostname from /etc/host first hostname and updating config files with DR hostname, This leads to failure of the application.
```

Reviewed By: jaymzh

Differential Revision: D5558640

fbshipit-source-id: 0910531af7e3087ed7aff1a8f1a21301fdbe06f5