Do not run grub2-install on efi systems
Summary:
Breakage on recently patched EFI systems is causing them to boot using the BIOS grub config
rather than the EFI config.  This is because we are running grub2-install upon upgrade of
the grub packages, about which https://fedoraproject.org/wiki/GRUB_2 has the following to
say:
	grub2-install shouldn't be used on EFI systems. The grub2-efi package installs a prebaked
	grubx64.efi on the EFI System partition, which looks for grub.cfg on the ESP in /EFI/fedora/
	whereas the grub2-install command creates a custom grubx64.efi, deletes the original installed
	one, and looks for grub.cfg in /boot/grub2/.

So, let's not.

Reviewed By: davide125

Differential Revision: D5841368

fbshipit-source-id: 2e4a99fa9dbff98f4edb277d24c64ce1246694ca