look in both pidfiles for rsyslog pid
Summary:

rsyslog 8.24 uses a new pidfile. We were overriding the logrotate config to use
the old pidfile. So after upgrading the package, the pidfile would change, and
when logrotate rotated the files it wouldn't SIGHUP the new daemon since it was
told to look at the old pidfile.

Update the logrotate config to look in both the old and new pidfiles for the
rsyslog pid.

Reviewed By: jaymzh

Differential Revision: D5846981

fbshipit-source-id: 7fb3615ceb7d4eee78756d6fd483fe62cef765c9