Add service_unit_options for setting [Unit] options
Summary:
This defines a `service_unit_options` key that allows you to specify options that should be set in the `[Unit]` section of the service unit. (as opposed to `service_options` which sets options in the `[Service]` section)

I'm making this change because I want to add a `Before=` ordering dependency for one of my services. If you don't like this change, I can stop using fb_timers and instead package the units with my RPM.

Reviewed By: zfjagann

Differential Revision: D5956732

fbshipit-source-id: 684e0914b8aa04207f6d0e652fbf1fa488489e39