Add "lock_file" option to take lock when converging
Summary:

Some mounts may be dynamically managed by other system processes. For example,
this may be done to temporarily change from `ro` mode to `rw` on a particular
mount to reduce exposure of critical mounts.

This diff allows fb_fstab to cooperate with such applications that use flock
semantics by requiring an exclusive hold of the cooperative lock for this
mount.

Reviewed By: patcox

Differential Revision: D6737725

fbshipit-source-id: ca03e4261286cc535e4b8d566a963a1e152fa280