consolidating and fixing fstab specs
Summary:
We've updated our spec runner to handle contexts correctly, and this
caused some failures in specs that have multiple contexts and stubbed library
helpers. The `before` block that stubs the library helper (e.g.
`:generate_base_fstab`) needs to be at the `recipe 'fb_fstab::default` level,
not within any context blocks, so it will be stubbed for all of the contexts.

In D6696691, phild actually fixed the runner and `bad_labels_spec` but the
fixed runner now causes `labels_spec` to fail on sandcastle (its actually
attempting to access and read the fstab on the sandcastle host and getting a
permission denied, P58913916, P58913921, and P58913919). Since we can actually
use contexts correctly now, we can consolidate these specs and make this
easier on us.

We could speed these tests up (specifically the uuid spec) by consolidating
some of the logic, e.g. pass in two mounts, one that passes an `only_if` and
one that doesn't and validate that the file renders correctly, but I'm not
doing that today.

Reviewed By: jaymzh

Differential Revision: D6755847

fbshipit-source-id: 41cabe72ae3d99585adc967a5e0c5bc27da0c8ff