break chef when iptables renders a bad config
Summary:
When adding a bad iptables rules to `fb_iptables` chef will render the `/etc/sysconfig/iptables` file and then will notify `execute[reload iptables]` which will fail to reload iptables.
The problem is that the following chef runs will not fail since `/etc/sysconfig/iptables` won't have any changes which means we will not notify `execute[reload iptables]`.

Of course this is the same for ipv6 I assume.

In order to solve this I'm using the `verify` feature of `template` to verify the iptables configuration before actually applying them.

Reviewed By: petermblair

Differential Revision: D7153691

fbshipit-source-id: c71c5fcc9408d894a8200856605f12b7e53d8b6f