fb_fstab: fix base-mounts-masking
Summary:
This is a quick-fix with minimal changes since it's a Friday evening. The right
fix is reasonably safe, but more invasive and I'm ont willing to do it on a
Friday evening.

This change simply allows a specific "base mount" (thing from Anaconda) to be
overidden by mount_point match *if* the base-mount is UUID and the
user-specified mount is LABEL. This is the only way to get out of the stuck-ness
of people who once used UUIDs and anaconda pre-scripts but now have moved to
Storage API and would like hotswap to work sanely.

Reviewed By: zfjagann

Differential Revision: D7314562

fbshipit-source-id: 22bf73e5412685d69c88eb08dade1d08e1d0d832