tmpreaper should use mtime by default
Summary:

`fb_tmpclean` uses `mtime` by default for tmpwatch (on CentOS)
but does not set a default when using tmpreaper (on Debian and macOS).

The default for tmpreaper is to use `atime`, which is a problem documented
on its manpage because when recursing on directories, testing their contents
would update the atime of the directories and so the directories themselves
do not get removed.

Switch the default to using `mtime` for consistent behavior, and make the
tmpreaper template honor the `timestamptype` property but still set
`--mtime-dir` when using `atime` so directories are cleaned up according to
their modification time regardless.

Reviewed By: joshuamiller01

Differential Revision: D7332112

fbshipit-source-id: f7f31d8c083655072f0eccb6512ac88e21139846