fb_dracut: rebuild the initramfs when /etc/sysctl.conf if updated
Summary:
By default dracut includes `/etc/sysctl.conf` in the initramfs. This can lead to subtle issues if a sysctl that was previously managed becomes unmanaged, as it will still be set at initramfs stage, but won't be explicitly unset later on. This forces a rebuild of the initramfs whenever the file changes, preventing the issue since the start.

Reviewed By: jaymzh

Differential Revision: D7345816

fbshipit-source-id: 3180c5b4da6a494e0d7c6a012b2db83d2a7d432b