Only save and restore iptables when necessary
Summary:
In order to prevent reloading the iptables modules unnecessarily, we
need to adjust how we dump and restore during chef runs and ensure we
don't spontaneously load empty ip6tables rules.

Saving and restoring empty tables loads the modules, so we need to
avoid that.  This change ensures we only perform the iptables
save/restore cycle if a table is active by checking
`/proc/net/ip6_tables_names`.

Reviewed By: jaymzh

Differential Revision: D7628961

fbshipit-source-id: 778c062650d1dbbdf0695387134228c1a6ed2c3e