move fb_fstab to custom resource
Summary:
I'm seeing a change in how we count updated resources and think it may be related to how chef 13 is referencing the resource. In chef 12 we would see any repeating resources that were occurring _inside_ the provider, not just the outer resource, but in chef 13 we're just getting a report of the outer resource and I'm wondering if this is due to the change/move to support custom resources instead of LWRP. Since we need to migrate to custom resources anyway, I thought I'd use this provider/resource (since it's a repeating one) as a good sample test case.

Example:
on a chef 13 host:
```
/var/chef/outputs/updated_resources.counts
{
  "version": "2",
  "resources": {
    "execute[/bin/bash /usr/facebook/configerator-client/bin/dir-setup.sh backup]": {
      "count": 21638
    },
    "fb_rsync[/usr/facebook/ops/scripts/]": {
      "count": 17
    },
    "fb_fstab[handle_mounts]": {
      "count": 17
    },
    "fb_rootcanal_tls_key[manage tls keys and certs]": {
      "count": 17
    }
  }
}
```
but on a chef 12 host:
```
{
  "version": "2",
  "resources": {
    "ohai[reload filesystems for fb_fstab]": {
      "count": 11404
    },
    "execute[/bin/bash /usr/facebook/configerator-client/bin/dir-setup.sh backup]": {
      "count": 11404
    },
    "execute[rootcanal-tls-mcrouter]": {
      "count": 11403
    }
  }
}
```
In the fb_fstab example I'm changing here, `ohai[reload filesystems for fb_fstab]` is an internal resource within the `fb_fstab[handle_mounts]` resource. It is not explicitly included in a recipe, but called as part of `fb_fstab[handle_mounts]` action. I have been able to verify that the `ohai[reload filesystems for fb_fstab]` is being executed in the chef 13 runs, so I may go spelunking in to the recipe dsl if this doesn't work to better understand how the data is reported back to us.

Reviewed By: jaymzh

Differential Revision: D8151019

fbshipit-source-id: 65de63d8f59bfefc7647386ed65fb286bdfee19e