rsync provider to use the rsync class method
Summary:
Continue to support rsync client creation outside of the custom resource, for the use cases where the custom resource doesn't support the options needed for the rsync.

To do this, I'm eliminating the Helpers module that I created to include in the resource's `action_class` and just calling the class method directly in the resource.

Reviewed By: jaymzh

Differential Revision: D8431394

fbshipit-source-id: a1b6e6dff0a5a635bb6a42e8d3fa52056e708195