Update fb_systemd to not restart systemd-logind on hosts which use target: graphical.target
Summary:

CPE would like to consume fb_systemd for the client fleet, however when fb_systemd performs a restart of systemd-logind the user looses keyboard and mouse control of the graphical interface.  This does not affect multi-user.target used on servers however since clients do use a graphical interface we need to gate this restart notification.

Looks similar to the following bug report

https://bugzilla.redhat.com/show_bug.cgi?id=1540321

Reviewed By: davide125

Differential Revision: D8960864

fbshipit-source-id: fba683bade96a4bd89f081f7485be7f685d88657