Add a chefignore
Summary:
I've primarily sourced the contents of the chefignore from the chefignore of chef's ohai cookbook. Some of the items in the chefignore (namely the berksfile items and such) are things that Facebook don't normally have in our cookbooks but may be included in cookbooks we use from upstream. Per Chef, this file needs to exist here:
> The chefignore file can be located in any subdirectory of a chef-repo: /, /cookbooks, /cookbooks/COOKBOOK_NAME/, roles, etc.

Reviewed By: davide125

Differential Revision: D9188977

fbshipit-source-id: 3d8b30c8985ca2a51c32a9f198a10bd90f23898a