prep to handle downgrade of rubocop from 0.52 to 0.49
Summary:
The new version of chefdk has a lower version of rubocop than what we were using previously. I need to temporarily drop the new options, do the cleanup, and then convert everyone but without users getting confused in the transition. I created a `.rubocop_49.yml` to define the behaviors and configure the lint to help transition to chefdk and allow me to roll back as needed. To help keep my sanity while I do this, I've done the following things in the `.rubocop_49.yml` but have not modified the existing `.rubocop.yml` to preserve the existing behavior:
* added the cop departments to each cop, helps me find the cop in the violation docs
* alphabetized and grouped the cops
* updated `.rubocop_49.yml` to reflect the cops and configuration options available to us in the version of rubocop included in chefdk
* added logic to `fb_rubocop` to pull in the correct `.rubocop.yml` based on if the user has been converted to chefdk or not

Reviewed By: jaymzh

Differential Revision: D10317371

fbshipit-source-id: 030cb3f047c90fc2b211e6096721165c923fb5b0