Update fb_iptables iptables_active? method to handle non existent files.
Summary:

fb_iptables fails with the following error in the context of the first chef run

```
Errno::ENOENT
    -------------
    No such file or directory @ rb_sysopen - /proc/net/ip6_tables_names
```

It appears the iptables_active? method can not handle when /proc/net/ip6_tables_names does not exist.

As part of the recipe it unloads and uninstalls firewalld, when the service is stopped ip6_tables_names is deleted however the ip_tables_names for ipv4 sticks around.

Reviewed By: davide125

Differential Revision: D10343256

fbshipit-source-id: ab614da7da76bb26d9ceb7070b4567bc3eaf144e