fix fb_swap to allow resizing swap
Summary: In a previous change, we finished a migration to add in the ability for encrypted swap. However, during the course of that migration setting the variable `swap_device` was dropped in favor of the new variable `base_swap_device` which is set by the new library. This particular execute block was not updated with the new variable, and I discovered this while trying to walk the nested run_contexts to build the resource collection.

Reviewed By: petermblair

Differential Revision: D13458511

fbshipit-source-id: b46e15655381e79302933bfbce2ed93a7f520549