Remove encryption support
Summary:

Direction is to use swap *files* where possible, which aren't block devices, and therefore can't be encrypted directly. Instead I think we should aim for full disk encryption so we get everything. This is being done to remove a lot of complexity and code surface area of code we weren't actively using.

Reviewed By: jaymzh

Differential Revision: D13583306

fbshipit-source-id: dd3276acc5cfec1ba73069ccc344deebac08d873