Add support for systemd timer
Summary:
Currently logrotate is run by cron. Modern systems come with systemd which allows to run cron like jobs.
The cron job comes with the RPM does not allow to use locking, so in theory it's possible that more than one job runs. There is an option to overwrite the cron job description with one which allows locking and manages priority with nice/ionice.
Systemd allows to achieve all of this in cleaner way.

In addition to this, the current code does not allow to switch back to the job description came with the RPM package.

This changes allows to use systemd timer by default, and if one wants to opt out, it will remove systemd stuff and install either the cron job with the locking if the `add_locking_to_logrotate` is true or the one from the CentOS 7 RPM otherwise.

Reviewed By: malmond77

Differential Revision: D13703282

fbshipit-source-id: 90cef40df932791737d3223bfb4f292c40a3dac7