Don't use systemd timer on containers
Summary: systemd daemon cannot be reloaded on a system where systemd is not an init system. We had guards for macos and centos6 which prevented systemd timer from being used on these system. However, if one runs this cookbook on a system which is not one of these, and the systemd is not an init, chef run fails. Check explicitly if the system is using the systemd as an init system.

Reviewed By: NaomiReeves

Differential Revision: D13773670

fbshipit-source-id: 37bfdfcf1fcc264a327388a8e1e5410bb7a7b5d9