Don't fail recipe if systemd daemon cannot be reloaded
Summary: If the unit exists on a system with no systemd configured as init, the recipe will remove it and try to reload the systemd daemon. It will fail on this system with "System has not been booted with systemd as init system (PID 1). Can't operate". Remove the file instead and do damon reload as a best effort.

Reviewed By: nikitakoshikov

Differential Revision: D13778453

fbshipit-source-id: 706dc484a5f90bfd23aa503036bad7f9bceb9f6e