Fix masking logic
Summary:

It turns out that my expression for masking swap wasn't the exact opposite of the bit that starts/enables it. I've written both so it's clear that it's predicated on swap being enabled at all, and that it'd have a positive size. It's strictly unneccessary in the first case, but it's easier to read now.

Reviewed By: htejun

Differential Revision: D13929349

fbshipit-source-id: 3517b486eea21fada523a6603a492bd4c00ace59