improve docs and tests around setting user
Summary:

Setting a user for a task to run as is non-obvious compared to the cron
cookbook.

This adds docs on how to do this, as well as some warnings if you pass unknown
keys. We also provide a forward breadcrumb if it looks like our client intended
to set a user.

Reviewed By: malmond77

Differential Revision: D13945602

fbshipit-source-id: 05323714273d7ce2a9a749a37e5913f53b23572d