fallback to node['filesystem'] if node['filesystem2'] is not available (removed in Chef 14 on Linux)
Summary:

On Linux and Mac, as of Chef 13, `node['filesystem']` and `node['filesystem2']` were identical and in Chef 14, `node['filesystem2']` is dropped.

For FreeBSD and other platforms, they become identical in late 15 and FS2 is dropped in late 16.

  https://docs.chef.io/deprecations_ohai_filesystem_v2.html

So we should always try FS2 and fall back to 1 (`node['filesystem2']` is not available, `node['filesystem']` is the new format).

Reviewed By: NaomiReeves

Differential Revision: D14353605

fbshipit-source-id: b471f0b27796880cfbdcd02cbd9e2a4b6bb3cf23