fb_iptables: Allow loading rules on tables not currently active
Summary:
In a previous change, an optimization was made to prevent the reloading of iptables modules of inactive tables. This is great for dynamic chains and cleaning up unused modules, but it prevented loading new rules to a table that we haven't previously activated.

In this diff I'm passing in static chains for the table where the user has defined a custom rule to `/etc/fb_iptables.conf` and adding a new value for `fb_iptables_reload` to check for when testing if a reload should occur. This test will only load modules for active tables or tables where we are configuring new rules.

Reviewed By: jaymzh

Differential Revision: D14235773

fbshipit-source-id: 58bef7b1d6220a344193dd3eeec08f33017c0e8d