add api for excluding paths for systemd-tmpfiles
Summary:

In some contexts the /sys filesystem is not writable, as a result systemd-tmpfiles will fail with exit code 73 if it attempts to process a tmpfiles.d config file with the "w" directive for a file in /sys.

One such example is /usr/lib/tmpfiles.d/selinux-policy.conf from the selinux-policy rpm.

Add generic support for adding path prefixes to be excluded by systemd-tmpfiles.

Reviewed By: malmond77, kevinpaulmccormick

Differential Revision: D14934534

fbshipit-source-id: 88ab645861b2756de8a550ddb286f0893d773062