Support fb_systemd_reload on non-booted systems
Summary:

It's common to define new unit files via `file`, `cookbook_file`, `systemd_unit` or `fb_systemd_override`. Once these are in place, it's common to notify `fb_systemd_reload[...]` to reload the daemon. Today if the system is not booted, then these notifies must be skipped.

Instead of forcing all resources that perform notifies to take this into account, the resource and resource provider itself can gracefully avoid a reload. This simplifies use of this cookbook.

Reviewed By: davide125

Differential Revision: D15084968

fbshipit-source-id: f6591d8db841d86e054f2c6f18bb8227a7164ca9