Use lsblk instead of direct sysfs interactions
Summary:

The code was passing the full pathname into the glob instead of the basic name, e.g. `/sys/block/*//dev/sda3`. Turns out that directly interacting with sysfs is non trivial. It's simpler to use lsblk to ask if a device (even a partition) is based on rotational media or not.

Reviewed By: htejun

Differential Revision: D15302086

fbshipit-source-id: 0c204832909504d135a5eaab318a194d64b965d0