Force dependency between timer and service
Summary:

Because of https://github.com/systemd/systemd/issues/6680
This is listed as the work-around, and we have plenty of hosts which seem to be experiencing this bug.
My previous attempt was wrong, put it in the right place this time so it works. It did work on the test host.

Reviewed By: dsesh, davide125

Differential Revision: D16135632

fbshipit-source-id: 70a20f27c7a08bd79d6293caef8b7693c21ad92f