support "reexec" option in reload custom resource
Summary:
Refactors the fb_systemd_reload stuff so it can be reused to implement reexec using the same code. The reexec is used during systemd upgrades.

This should be done as part of the RPM installation but there is no built in way to `systemctl daemon-reexec` for all running user managers (upstream issue filed https://github.com/systemd/systemd/issues/13065).

Reviewed By: NaomiReeves

Differential Revision: D16274280

fbshipit-source-id: c557a9bf2c5d9502d8e926d6616f24bfca960dd9