codemod node method macosx? to macos?
Summary: In a previous change we created an alias for `node.macosx?` to `node.macos?` and apparently foodcritic doesn't follow alias's for methods or we should have used an `alias_method` instead of `alias`. Once I've isolated if this is an issue upstream or with our custom foodcritic rule, I'll take the correct action to fix the actual issue, but for now I'm fixing the codebase.

Reviewed By: joshuamiller01

Differential Revision: D16295510

fbshipit-source-id: 203f0216a16c7be454ac1101bbd5abf710c3c24f