fb_systemd: suppress extra new line in to_ini
Summary:
When comparing differences between sandcastle's custom images and the first run, I discovered that systemd config files were changed to add an additional newline character.

It appears that the sharding is such that the custom images use the template code, while at task runtime it uses the to_ini method.

It's unclear to me that we still need both implementations, but at least for now this diff simply tweaks the template to not append an additional newline in the to_ini branch of the conditional so the output of both should be identical.

Reviewed By: joshuamiller01

Differential Revision: D16814891

fbshipit-source-id: bd48fb2325ed9ef472c18ecd42f65743ff16d7d9