New cookbook: fb_ssh
This is a cookbook to manage SSH using the FB attribute-driven model.

It handles daemon and client configs as well as authorized_keys and
authorized_principals.

There are few things worth noting here: The API for public keys uses a
`data_bag`. Grocery delivery and taste-tester have long supported
data_bags even though FB doesn't use them in prod... but it doesn't
matter since FB uses principals in prod and not keys, so this shouldn't
be an issue no matter what.

Signed-off-by: Phil Dibowitz <phil@ipom.com>