fb_systemd: add spaces before and after "=" in systemd.conf.erb template
Summary:
Depending on shard, the systemd.conf.erb template uses the FB::Systemd.to_ini function or normal template expansion.  While investigating the cause of unexpected changes to systemd config files, I discovered that the "to_ini" method adds spaces before and after the "=" in assignments while the template expansion does not.

It's unclear to me that we still need both implementations, but at least for now this diff simply tweaks the template to add spaces around the "=" in the template expansion to match the output of the to_ini branch of the shard conditional.

Reviewed By: nikitakoshikov

Differential Revision: D16836339

fbshipit-source-id: d1b1f3dfa484e4e24b0ba06c2fbe77a65cd9e080