Restore old behavior of the device_name_from_partition
Summary:

The regression came with the D17499312, there were 2 methods doing very the same things, and they were deduplicated, but it turned out, that the behaviour of the used one was incorrect in some cases (for /dev/mdX).

Ran the script with the old and new methods: P112440913
It works correctly for the nvme, but does not for md:
$ ruby bin/zxc.rb
/dev/md0
/dev/md
/dev/nvme3n1
/dev/nvme3n1

So, I restored the code from the removed method to the deduplicated one.

Reviewed By: davide125

Differential Revision: D17518088

fbshipit-source-id: 6682ddc04dc29361abeed1854dc7956431f94b58