Run systemd timers under an explicitly named slice 0->1%
Summary:
This runs systemd timers under the cgroup
`system.slice/system-timers.slice/system-timers-<timer_name>.slice/<timer_name>.service`

In doing this we can attribute resource consumption of all timers and
individual timers as well because the slices persist.

If this method is agreed upon, we can just shard this.

Reviewed By: joshuamiller01

Differential Revision: D17397380

fbshipit-source-id: 7555d24cdab8198ad724917d87f329672a8518e1