new cookbook: fb_sssd
This is a very simple cookbook to manage `sssd` a common staple on
modern Linux machines - both servers and client machines.

I've tried to stick as close as possible to my recollection of the
internal FB cookbook. Other than some extra symlinks and a bit of
internal config, I think this should be a nearly drop-in replacement
that will be easy to migrate to.