Test all cookbooks
Currently only cookbooks referenced in the sample init are tested,
but things like apache and friends that are likely to be included
from a "tier" role/cookbook and not from init are never tested.

This adds an entry point in testing to include those so they get
tested as well.