fb_apache: various fixes (#85)
Summary:

* Handle multiple vhosts of the same "name"

It's common to have various `<VirtualHost *:80>` entries, but the API
doesn't allow that. I took a page out of the Chef resource book and
tweaked the API to allow you to use a different name, but then set
the name of the virtual host with a `_virtualhost` key.

* Handle rewrite rules with no `RewriteCondition` clause

Currently if there's no `conditions` array in a `rewrite` entry the
template will crash, this fixes that.

* Handle the confdir on debian/ubuntu

Closes https://github.com/facebook/chef-cookbooks/issues/84

* Don't include built-in modules on Debian/Ubuntu

Modules that are built in will cause apache to fail startup if you list
them as modules to load.

* Cleanup MD in the README.md
Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/85

Reviewed By: claudiozz

Differential Revision: D18277426

Pulled By: davide125

fbshipit-source-id: 03b208921a06fec352bbeffb69b97659f2e37c7c