Test all cookbooks (#75)
Summary:
Currently only cookbooks referenced in the sample init are tested,
but things like apache and friends that are likely to be included
from a "tier" role/cookbook and not from init are never tested.

This adds an entry point in testing to include those so they get
tested as well.

NOTE: This is based ON https://github.com/facebook/chef-cookbooks/issues/77 so both changes are in it. Will rebase after that is landed (if it works)
Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/75

Test Plan:

Reviewed By: NaomiReeves, malmond77

Differential Revision: D18277500

Pulled By: davide125

fbshipit-source-id: 57057938bb2545a38ee3d1e6afe8c237e7804577