fb_apt: significant speedup; os upgrades
This PR changes `fb_apt` from running a `dpkg -S` on every single
keyring, and instead runs a single one. This is significantly faster,
saving ~3 seconds on a machine with only a few keyrings.

This also allows the user to override the distro which is necessary
for instrumenting OS upgrades across distros.

Signed-off-by: Phil Dibowitz <phil@ipom.com>