fb_apt: significant speedup; os upgrades (#89)
Summary:
This PR changes `fb_apt` from running a `dpkg -S` on every single
keyring, and instead runs a single one. This is significantly faster,
saving ~3 seconds on a machine with only a few keyrings.

This also allows the user to override the distro which is necessary
for instrumenting OS upgrades across distros.
Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/89

Test Plan:

Reviewed By: malmond77

Differential Revision: D18456049

Pulled By: davide125

fbshipit-source-id: daa35504e8fea1d6d7f87236de3976a7c11a639b