drop v52 rubocop.yml now that everything is on chefdk
Summary: We made a temporary rubocop config, `.rubocop_49.yml` to help with the downgrade of `rubocop` to v49 from v52 during our migration to chefdk. The migration to chefdk is complete and now I can rename the `.rubocop_49.yml` (used with v49) to `.rubocop.yml` and drop the old v52 config, the current `.rubocop.yml`, in order to prepare for the next version of chefdk.

Reviewed By: petermblair

Differential Revision: D18663596

fbshipit-source-id: db8effd92429167b6307f2f4a7349b80dfe11e6a