Add support for btrfs sends stream
Summary:

This adds support to tupperware agent to handle both subvolumes as they are created now and with sendstreams.

This is complementary to the installer change.

This was _incredibly_ painful to test, mostly due to deficencies in how we test provisioning and chef changes, so if you've been wondering what did claudioz@ do with sendstreams? It was mostly babysitting cyborg jobs and making chef/provisioning changes _just_ to test stuff like this without going mad.

This was horrible and I hope to improvie it (some work has been already put in) or never have to do it again in my lifetime.

Reviewed By: joshuamiller01

Differential Revision: D19102184

fbshipit-source-id: 7e28e98b1bd2d49a633e75bf5f5b173635fc5207