allow chrony to use up to 10MiB or RAM for client status
Summary:
https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/system_administrators_guide/sect-hw_timestamping#s2-enabling-interleaved-mode

According to this paper this is essencial part for hwtimstamps to function.

https://chrony.tuxfamily.org/doc/3.4/chrony.conf.html

```
clientloglimit limit
This directive specifies the maximum amount of memory that chronyd is allowed to allocate for logging of client accesses and the state that chronyd as an NTP server needs to support the interleaved mode for its clients. The default limit is 524288 bytes, which is sufficient for monitoring about four thousand clients at the same time.
```
This is needed for hardware timestamps to function better on high loaded clusters.

Reviewed By: nikitakoshikov

Differential Revision: D20392630

fbshipit-source-id: eb1b35a7ffa14a428f75cb604611b07de577a99b