filesystem2 => filesystem conversion: fb_storage
Summary:
Converting the `fb_storage` cookbook over to the `filesystem_data` node method.

In chef 13, the `filesystem` ohai plugin was replaced with the `filesystem2` plugin, and the `filesystem2` data is written to both the `filesystem2` and `filesystem` apis. All uses of the old `filesystem` api were converted to the `filesystem2` api before I upgraded the fleet to chef 13 (tracked in T28228267). In order to upgrade to Chef 14+, we need to remove the references to the `filesystem2` api. However, there are some outlier operating systems that missed the conversion and so we're going to use a node method to detect what api is available and use that one so we can continue to support this transition for those other platforms.

official deprecation: https://docs.chef.io/deprecations_ohai_filesystem.html

PR to covert BSD: https://github.com/chef/ohai/pull/1181
PR to convert the rest of Unix (solaris, aix): https://github.com/chef/ohai/pull/1266
PR to convert Windows: https://github.com/chef/ohai/pull/1267

Reviewed By: davide125

Differential Revision: D20634208

fbshipit-source-id: 90d3d3a00763d4e74fbda648c57a0abd6b649656