Define a new method for rpm versions
Summary:

Finding the version of an installed RPM is an extremely common need. I've seen this kind of code littered all over the cookbooks. It was sort of okay for CentOS 7, but now we have `dnf` in the mix, we are repeating the code a lot. We can use this to reduce repetition.

Reviewed By: nikitakoshikov

Differential Revision: D20922382

fbshipit-source-id: b383cfa888eb37153c04c2eb0c97040e5a8609eb